[BEAM-480] Move insertAll() from BigQueryTableInserter to BigQueryServices
